Rename files from index to CUI header and fix imports
