+++ src/components/CUIHeader/CUIHeader.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 
-import styles from './index.module.scss';
+import styles from './CUIHeader.module.scss';
 
 const cx = classNames.bind(styles);
