+++ src/components/CUIHeader/CUIHeader.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import CUIHeader from './index';
+import CUIHeader from './CUIHeader';
 
 describe('CUIHeader', () => {
   it('Displays Controlled Unclassified Information', () => {
