+++ src/scenes/SystemAdmin/Home.jsx
@@ -4,7 +4,7 @@ import { Admin, AppBar, fetchUtils, Layout, Resource } from 'react-admin';
 import { createBrowserHistory } from 'history';
 import React from 'react';
 import Menu from './shared/Menu';
-import CUIHeader from 'components/CUIHeader';
+import CUIHeader from 'components/CUIHeader/CUIHeader';
 import UploadShow from './Uploads/UploadShow';
 import OfficeUserList from 'pages/Admin/OfficeUsers/OfficeUserList';
 import OfficeUserShow from 'pages/Admin/OfficeUsers/OfficeUserShow';
